Added CorsFilter to enable secure cross site scripting of VTN
See: http://tomcat.apache.org/tomcat-7.0-doc/config/filter.html#CORS_Filter
And: http://en.wikipedia.org/wiki/Cross-origin_resource_sharing
This is done to allow a web page using javascript to be able to make calls
to our REST APIs even though it does not originate in our domain.

Note: This was done first for VTN because the user asking for the capability
on IRC was working on building a web page to program some VTN stuff.

PLEASE DO NOT MERGE THIS WITHOUT THOUGHT AND DISCUSSION.
This patch was done to get something out there for for routerpro
to try out... its not much tested yet.

Signed-off-by: Ed Warnicke <eaw@cisco.com>