Update patch set 1
Patch Set 1:

Do the services mean IPC service? If so, it is alright to implement a module not to provide a service.
Basically, a program which provides some functionalities on the pfcd clone should be implemented as a module. A library implementation is used for a framework like a IPC. Or it is used for a library which is shared between a normal process and the pfcd clone.
Can I ask you to convert this library to a module?

Patch-set: 1
Reviewer: Gerrit User 245 <245@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0