Update patch set 1
Patch Set 1:

I think you should change namespaces of some classes.

In the patch, you've created new namespaces such as "restjson" and "odc", and put some cleasses into the nemespaces.
Are there any particular reasons for this?

All existing classes running in modules of VTN Coordinator are in the "unc" namespace.
So it is reasonable to me to change namespaces of classes in the patch to the same namespace "unc".

Patch-set: 1
Reviewer: Gerrit User 396 <396@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0