Update patch set 2
Patch Set 2:

Could you change the commit message?

I think that "drvvtnd" does not exist at the master branch of VTN Git repository.
Therefore, the commit message of this patch set is a little confusing.
For example, the commit message of the following abandoned change is better for the commit message of this patch.

  https://git.opendaylight.org/gerrit/#/c/1657/

Patch-set: 2
Label: Verified=0