Fixed VTN coordinator build failure with the latest json-c library.
The name of json-c library was changed to libjson-c as of version 0.11.
This patch detects compiler and linker flags required to link json-c library,
and defines them as makefile variables.

If you want to use libjson-c in your component, put the following lines
to your Makefile:

  EXTRA_CPPFLAGS          = $(JSON_C_CPPFLAGS)
  EXTRA_LIBDIRS           = $(JSON_C_LIBDIRS)
  EXTRA_LDLIBS            = $(JSON_C_LDFLAGS)
  EXTRA_RUNTIME_DIR       = $(JSON_C_RUNPATH)

In addition, libcurl is also detected by configure script.
The following lines need to be added to your Makefile to link libcurl:

  EXTRA_CPPFLAGS          = $(LIBCURL_CPPFLAGS)
  EXTRA_LIBDIRS           = $(LIBCURL_LIBDIRS)
  EXTRA_LDLIBS            = $(LIBCURL_LDFLAGS)
  EXTRA_RUNTIME_DIR       = $(LIBCURL_RUNPATH)

Note that this patch requires pkg-config command in build host.

Change-Id: Ie1a0352c7db598160269c6045b3ede04e57364eb
Signed-off-by: Shigeru Yasuda <s-yasuda@da.jp.nec.com>