Initial Feature File for VTN Project
This feature file is yet to be tested and must be revisited after
both OVSDB versions and features along with controller NSF features are
properly merged in however it does define the proper basic structure for
it.

Get in touch for any question about it.. Tests will be coming next week.

Change-Id: Ia245d8919b8d20e713d3d7f88764c16dd7a511d8
Signed-off-by: Mathieu Lemay <mlemay@inocybe.com>