Disable integration tests temporarily.
VTN Manager's IT is being broken by the data change listener issue
tracked by Bug 3344. We decided to disable IT so that we can merge
patches. IT will be enabled after that issue is fixed.

In addition, this patch disables VSEM provider build.
VSEM provider build requires nuget command, but it is sometimes
missing in Jenkins VM.

Change-Id: Ifd20c45cb7dbfd598488d8e3d8353f6c1d20d06b
Signed-off-by: Shigeru Yasuda <s-yasuda@da.jp.nec.com>