Bug 3818: Restore switch link from network-topology on link-up.
Topology-manager does not remove link information when link-down is
detected on a port. So a data change event may not be raised if link state
is changed twice or more in a very short term.
So the VTN manager needs to restore inter-switch link information from
network-topology when link-up is detected on a port.

Change-Id: I93a8a8080f27a514c542b705e57649eeca72196f
Signed-off-by: Shigeru Yasuda <s-yasuda@da.jp.nec.com>
(cherry picked from commit 24d92f649718439d1c81dd66a71564f373e2ee5e)