Update patch set 1
Patch Set 1:

> Alright. Another possibility is below, but I feel that may be too
 > un-intuitive.
 > 1. mvn install -DwithoutFeatures=true
 > 2. mvn install -DwithoutNonFeatures=true

I don't think this is a good solution as it's kind of hackish and unintuitive for the developer. The best solution in my opinion is to fix yang-maven-plugin so that it doesn't cause re-compiles.

The default "mvn clean install" should always work for a developer in my opinion.

Patch-set: 1