Update patch set 1
Patch Set 1: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/vtn-validate-autorelease-beryllium/218/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/vtn-validate-autorelease-beryllium/218/ )

https://jenkins.opendaylight.org/releng/job/vtn-verify-beryllium-mvn33-openjdk8/5/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/vtn-verify-beryllium-mvn33-openjdk8/5/ )

https://jenkins.opendaylight.org/releng/job/vtn-distribution-check-beryllium/86/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/vtn-distribution-check-beryllium/86/ )

https://jenkins.opendaylight.org/releng/job/vtn-verify-beryllium-mvn33-openjdk7/5/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/vtn-verify-beryllium-mvn33-openjdk7/5/ )

Patch-set: 1