Bug 6543: Changed the default value of port-name
This patch changed the default value of the parameter "port-name" for
the odl-vtn-manager-neutron feature to "eth1" in order to make the CSIT
for the feature work well. As of now, the parameter is not configurable,
and the default value is always used. However, in the ODL CSIT
environment, this parameter needs to be "eth1". So this change is a
tentative fix. Ideally, the parameter "port-name" should be
configurable.

Change-Id: Id059618d975dc4982d498c7ab931c55026ba61bb
Signed-off-by: gvrangan <venkatrangang@hcl.com>