Fix Bug 7969, Enable using latest postgres  libraries
Postgres has fixed attribute validation, casuing  some VTN operations
to fail. This fixes the line of code  which adds port_id to the where clause of a query
when the value is empty. Without the fix the variable bindings with the query
adds values for switch_id and port_id, whereas  the where clause adds  only  port_id.
This results in  postgres  looking at  bindings for the value of port_id and instead  reads  the switch_id
which is of a different data type  and reports error.

Change-Id: Ic444a99fc1d2c51cc758f8dba745affb3a0b57d0
Signed-off-by: gvrangan <venkatrangang@hcl.com>