Update patch set 1
Patch Set 1:

Thank you for submitting the patch. I know this is a wip patch, but let me give you a comment.

I think we should write a quick comment in the description section about a Maven project which the POM models. That should not be a description about VTN project, but should be about an each Maven project. So all descriptions in POM files in the VTN project should be not the same, though all descriptions in your change have the same description as of now.

Patch-set: 1
Label: Verified=0